requirements.txt
imageio==2.9.0
lmdb==1.4.0
lpips==0.1.4
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.6.0.66
Pillow==9.2.0
pytorch-wavelets==1.3.0
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.11.1
torch==1.12.0
torchaudio==0.12.0
torchmetrics==0.9.3
torchvision==0.13.0
tqdm==4.64.0

